use sync resource detectors (#112)
*Issue #, if available:*

*Description of changes:*

Due to a
[bug](https://github.com/open-telemetry/opentelemetry-js/pull/4687) in
upstream, the `detectResourcesSync` doesn't wait for the async
attributes. This happens if we use the async version of some resource
detector. The
[fix](https://github.com/open-telemetry/opentelemetry-js/pull/4687) is
in another upstream release. We don't want to upgrade the upstream
version for now. To avoid this issue, we switch to the sync version of
the detectors and this should have no impact for the functionalities (as
`detectResourcesSync` waits for all async detectors anyway and works for
both sync and async detectors). In the future, even if we upgrade
upstream version, we don't need to switch back to async detectors.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.