Fix: Deprecated GitHub Artifacts (#152)
*Issue #, if available:*
Github workflows are failing due to:
```
Error: This request has been automatically failed because it uses a deprecated version of `actions/upload-artifact: v3`. Learn more: https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/
```

A downstream effect of upgrading to `v4` is this version no longer
allows uploading artifacts with same name.
See:
https://github.com/aws-observability/aws-otel-java-instrumentation/pull/1014/files

*Description of changes:*
We apply the same fix in the above PR


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.