SigV4 Authentication support for http/protobuf exporter (#156)
*Issue #, if available:*
Adding SigV4 Authentication extension for Exporting traces to OTLP XRay
endpoint without needing to explictily install the collector.

*Description of changes:*
Added a new class that extends upstream's OTLP http/protobuf span
exporter. Overrides the export method so that if the endpoint is CW, we
add an extra step of injecting SigV4 authentication to the headers.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.