Remove old SigV4 headers before signing (#159)
*Description of changes:*
There was a bug where the old headers were not cleaned before the
request was signed causing 403 errors to occur when exporting traces to
XRay OTLP endpoint.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.