Improve Lambda Layer cold start time #2 (#171)
*Description of changes:*
This PR is a rehash of #163 with some minor changes reverted after
several autoinstrumentation package dependencies were upgraded in #168.
The changes suggested by the unaddressed comments in the previous PR are
addressed here.

The primary change made to reduce the cold start time was to bundle the
ADOT autoinstrumentation package with webpack instead of using the TS
compiler. Using the same build targets as the upstream (see [this
PR](https://github.com/open-telemetry/opentelemetry-lambda/pull/1679/files)),
the cold start performance of the layer was improved by about 50%, from
~765ms to ~385ms.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.