Send main build metrics (#240)
*Issue #, if available:*

*Description of changes:*
Emit a failure metric if main build fails. Since this workflow is
triggered with pushes to main or a release branch, we want to be
notified if there is a failure with the build process or e2e tests.

Tested by temporarily adding an `on: push:` trigger to my own branch in
this repo and testing the updated workflow. Verified that failure metric
was successfully published to cloudwatch.

https://github.com/aws-observability/aws-otel-js-instrumentation/actions/runs/16764243839

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.