Fix issue where UDP Exporter throws error in async callback, which isn't caught (#289)
*Issue #, if available:*
In UDP Exporter, when sending via the socket, our callback method will
rethrow the error if there an error from the `send`. However, this is
done asynchronously, so it isn't actually caught in the surrounding
`try/catch`.

This `err` can happen when the UDP Packet size limit is reached. When
this error is thrown, it can cause the following error in Lambda:
```
[ERROR] [1761253646458] LAMBDA_RUNTIME Failed to post handler success response. Http response code: 403. 
{
    "errorMessage": "State transition from RuntimeResponseSentState to InvocationErrorResponse failed for runtime. Error: State transition is not allowed",
    "errorType": "InvalidStateTransition"
}
```

*Description of changes:*
- reject error, as opposed to throwing error because it cannot be caught
in the existing surrounding `try/catch`.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.