Set up trusted publishing for ADOT SDK and UDP exporter (#300)
*Issue #, if available:*

*Description of changes:*
Recent [npm security
changes](https://github.blog/changelog/2025-09-29-strengthening-npm-security-important-changes-to-authentication-and-token-management/)
mean that all tokens used for publishing to npm are now granular and
have a maximum lifespan of 90 days.

This PR migrates our Github workflows to use trusted publishing
following [npm's guidance
docs](https://docs.npmjs.com/trusted-publishers#configuring-trusted-publishing),
no longer relying on tokens. For the SDK release workflow, the npm
publish step is moved to the first publish step so that, in case of
publish failure, we will not be in a partial release state.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.